dispatcher/requirements.txt
fastapi
uvicorn
redis
httpx
pytest
pytest-asyncio
python-jose[cryptography]
